+++ flinkboot-test/src/main/java/module-info.java
@@ -1,5 +1,6 @@
 module io.github.sekelenao.flinkboot.test {
     requires transitive io.github.sekelenao.flinkboot.core;
+    requires transitive flink.streaming.java;
     requires transitive org.junit.jupiter.api;
     requires org.junit.jupiter.params;